// src/app/models/exchange/MarginSummaryForExchangeOrder.ts
export class MarginSummary {
    exchangeFees: number = 0;
    overhaulPrice: number = 0;
    otherCharges: number = 0;
    totalEstRevenue: number = 0;
    cogsFees: number = 0;
    overhaulCost: number = 0;
    exchangeOrderMarginSummaryId: number;
    exchangeQuoteId: number;
    otherCost: number = 0;
    marginAmount:number=0;
    marginPercentage:number=0;
    totalEstCost:number=0;
}